move sysinfo.go funcs to util.go
